requirements.txt
click==6.6
Flask==0.11.1
github-webhook==1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyGithub==1.29
PyYAML==3.12
requests==2.12.3
six==1.10.0
slackclient==1.0.2
websocket-client==0.40.0
Werkzeug==0.11.11
